El CRUD de asesorias se completo

+++ backend/controllers/products.controller.php
@@ -32,26 +32,23 @@ class ControllerProducts{
 	
 	}
 
-    public static function updateProduct($datos){
+	/*=============================================
+	CREAR PRODUCTOS
+	=============================================*/
 
-		if(isset($datos["idAsesoria"])){
+	public static function crearAsesoria($datos){
 
-			if(preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $datos["tituloAsesoria"])  && preg_match('/^[,\\.\\a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["descripcionAsesoria"]) ){
+		if(isset($datos["tituloAsesoria"])){
 
+			if(preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $datos["tituloAsesoria"]) && preg_match('/^[,\\.\\a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["descripcionAsesoria"]) ){
 
 				/*=============================================
 				VALIDAR IMAGEN PORTADA
 				=============================================*/
 
-				$rutaPortada = "../".$datos["antiguaFotoPortada"];
+				$rutaPortada = "../views/img/open_graph/default/default.jpg";
 
 				if(isset($datos["fotoPortada"]["tmp_name"]) && !empty($datos["fotoPortada"]["tmp_name"])){
-
-					/*=============================================
-					BORRAMOS ANTIGUA FOTO PORTADA
-					=============================================*/
-
-					unlink("../".$datos["antiguaFotoPortada"]);
 
 					/*=============================================
 					DEFINIMOS LAS MEDIDAS
@@ -75,7 +72,7 @@ class ControllerProducts{
 
 						$aleatorio = mt_rand(100,999);
 
-						$rutaPortada = "../vistas/img/cabeceras/".$datos["rutaProducto"].".jpg";
+						$rutaPortada = "../views/img/open_graph/default/".$datos["rutaAsesoria"].".jpg";
 
 						$origen = imagecreatefromjpeg($datos["fotoPortada"]["tmp_name"]);						
 						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
@@ -94,7 +91,236 @@ class ControllerProducts{
 
 						$aleatorio = mt_rand(100,999);
 
-						$rutaPortada = "../vistas/img/cabeceras/".$datos["rutaProducto"].".png";
+						$rutaPortada = "../views/img/open_graph/default/".$datos["rutaAsesoria"].".png";
+
+						$origen = imagecreatefrompng($datos["fotoPortada"]["tmp_name"]);						
+
+						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+
+						imagealphablending($destino, FALSE);
+				
+						imagesavealpha($destino, TRUE);
+
+						imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+
+						imagepng($destino, $rutaPortada);
+
+					}
+
+				}
+
+				/*=============================================
+				VALIDAR IMAGEN PRINCIPAL
+				=============================================*/
+
+				$rutaFotoPrincipal = "../views/img/categories/default/default.jpg";
+
+				if(isset($datos["fotoPrincipal"]["tmp_name"]) && !empty($datos["fotoPrincipal"]["tmp_name"])){
+
+					/*=============================================
+					DEFINIMOS LAS MEDIDAS
+					=============================================*/
+
+					list($ancho, $alto) = getimagesize($datos["fotoPrincipal"]["tmp_name"]);	
+
+					$nuevoAncho = 450;
+					$nuevoAlto = 450;
+
+					/*=============================================
+					DE ACUERDO AL TIPO DE IMAGEN APLICAMOS LAS FUNCIONES POR DEFECTO DE PHP
+					=============================================*/
+
+					if($datos["fotoPrincipal"]["type"] == "image/jpeg"){
+
+						/*=============================================
+						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
+						=============================================*/
+
+						$aleatorio = mt_rand(100,999);
+
+						$rutaFotoPrincipal = "../views/img/categories/asesorias/".$datos["rutaAsesoria"].".jpg";
+
+						$origen = imagecreatefromjpeg($datos["fotoPrincipal"]["tmp_name"]);						
+
+						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+
+						imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+
+						imagejpeg($destino, $rutaFotoPrincipal);
+
+					}
+
+					if($datos["fotoPrincipal"]["type"] == "image/png"){
+
+						/*=============================================
+						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
+						=============================================*/
+
+						$aleatorio = mt_rand(100,999);
+
+						$rutaFotoPrincipal = "../views/img/categories/asesorias/".$datos["rutaAsesoria"].".png";
+
+						$origen = imagecreatefrompng($datos["fotoPrincipal"]["tmp_name"]);						
+
+						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+
+						imagealphablending($destino, FALSE);
+				
+						imagesavealpha($destino, TRUE);
+
+						imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+
+						imagepng($destino, $rutaFotoPrincipal);
+
+					}
+
+				}
+
+				/*=============================================
+				PREGUNTAMOS SI VIENE OFERTE EN CAMINO
+				=============================================*/
+
+				if($datos["selActivarOferta"] == "oferta"){
+
+					$datosAsesoria = array(
+						   "titulo"=>$datos["tituloAsesoria"],
+						   "idCategoria"=>2,
+						   "detalles"=>$datos["detalles"],		
+						   "hour"=>$datos["hour"],				 
+						   "ruta"=>$datos["rutaAsesoria"],
+						   "estado"=> 1,
+						   "titular"=> substr($datos["descripcionAsesoria"], 0, 225)."...",
+						   "descripcion"=> $datos["descripcionAsesoria"],
+						   "keywords"=> $datos["pClavesAsesoria"],
+						   "precio"=> $datos["precio"],				
+						   "imgPortada"=>substr($rutaPortada,3),
+						   "imgFotoPrincipal"=>substr($rutaFotoPrincipal,3),
+						   "oferta"=>1,
+						   "precioOferta"=>$datos["precioOferta"],
+						   "descuentoOferta"=>$datos["descuentoOferta"],
+					   );
+
+				}else{
+
+					$datosAsesoria = array(
+						   "titulo"=>$datos["tituloAsesoria"],
+						   "idCategoria"=>2,
+						   "detalles"=>$datos["detalles"],	
+						   "hour"=>$datos["hour"],					
+						   "ruta"=>$datos["rutaAsesoria"],
+						   "estado"=> 1,
+						   "titular"=> substr($datos["descripcionAsesoria"], 0, 225)."...",
+						   "descripcion"=> $datos["descripcionAsesoria"],
+						   "keywords"=> $datos["pClavesAsesoria"],
+						   "precio"=> $datos["precio"],					  
+						   "imgPortada"=>substr($rutaPortada,3),
+						   "imgFotoPrincipal"=>substr($rutaFotoPrincipal,3),
+						   "oferta"=>0,
+						   "precioOferta"=>0,
+						   "descuentoOferta"=>0,
+						  
+					   );
+
+				}
+
+				ModelOpenGraph::addOpenGraph("open_graph", $datosAsesoria);
+
+				$respuesta = ModelProducts::crearAsesoria("products", $datosAsesoria);
+
+				return $respuesta;
+				
+
+			}else{
+
+					echo'<script>
+
+					swal({
+						  type: "error",
+						  title: "El nombre de la asesoría no puede ir vacía o llevar caracteres especiales",
+						  showConfirmButton: true,
+						  confirmButtonText: "Cerrar"
+						  }).then(function(result){
+							if (result.value) {
+
+							window.location = "asesorias";
+
+							}
+						})
+
+			  	</script>';
+
+			}
+		
+		}
+
+	}
+
+    public static function updateProduct($datos){
+
+		if(isset($datos["idAsesoria"])){
+
+			if(preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $datos["tituloAsesoria"])  && preg_match('/^[,\\.\\a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["descripcionAsesoria"]) ){
+
+
+				/*=============================================
+				VALIDAR IMAGEN PORTADA
+				=============================================*/
+
+				$rutaPortada = "../".$datos["antiguaFotoPortada"];
+
+				if(isset($datos["fotoPortada"]["tmp_name"]) && !empty($datos["fotoPortada"]["tmp_name"])){
+
+					/*=============================================
+					BORRAMOS ANTIGUA FOTO PORTADA
+					=============================================*/
+
+					if($datos["antiguaFotoPortada"] != "" && $datos["antiguaFotoPortada"] != "views/img/open_graph/default/default.jpg"){
+
+						unlink("../".$datos["antiguaFotoPortada"]);	
+		
+					}
+
+					/*=============================================
+					DEFINIMOS LAS MEDIDAS
+					=============================================*/
+
+					list($ancho, $alto) = getimagesize($datos["fotoPortada"]["tmp_name"]);	
+
+					$nuevoAncho = 1280;
+					$nuevoAlto = 720;
+
+					/*=============================================
+					DE ACUERDO AL TIPO DE IMAGEN APLICAMOS LAS FUNCIONES POR DEFECTO DE PHP
+					=============================================*/
+
+					if($datos["fotoPortada"]["type"] == "image/jpeg"){
+
+						/*=============================================
+						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
+						=============================================*/
+
+						$aleatorio = mt_rand(100,999);
+
+						$rutaPortada = "../views/img/open_graph/".$datos["rutaAsesoria"].".jpg";
+
+						$origen = imagecreatefromjpeg($datos["fotoPortada"]["tmp_name"]);						
+						$destino = imagecreatetruecolor($nuevoAncho, $nuevoAlto);
+
+						imagecopyresized($destino, $origen, 0, 0, 0, 0, $nuevoAncho, $nuevoAlto, $ancho, $alto);
+
+						imagejpeg($destino, $rutaPortada);
+
+					}
+
+					if($datos["fotoPortada"]["type"] == "image/png"){
+
+						/*=============================================
+						GUARDAMOS LA IMAGEN EN EL DIRECTORIO
+						=============================================*/
+
+						$aleatorio = mt_rand(100,999);
+
+						$rutaPortada = "../views/img/open_graph/".$datos["rutaAsesoria"].".png";
 
 						$origen = imagecreatefrompng($datos["fotoPortada"]["tmp_name"]);						
 
@@ -124,7 +350,11 @@ class ControllerProducts{
 					BORRAMOS ANTIGUA FOTO PRINCIPAL
 					=============================================*/
 
-					unlink("../".$datos["antiguaFotoPrincipal"]);
+					if($datos["antiguaFotoPrincipal"] != "" && $datos["antiguaFotoPrincipal"] != "views/img/categories/default/default.jpg"){
+
+						unlink("../".$datos["antiguaFotoPrincipal"]);
+		
+					}
 
 					/*=============================================
 					DEFINIMOS LAS MEDIDAS
@@ -132,7 +362,7 @@ class ControllerProducts{
 
 					list($ancho, $alto) = getimagesize($datos["fotoPrincipal"]["tmp_name"]);	
 
-					$nuevoAncho = 400;
+					$nuevoAncho = 450;
 					$nuevoAlto = 450;
 
 
@@ -148,7 +378,7 @@ class ControllerProducts{
 
 						$aleatorio = mt_rand(100,999);
 
-						$rutaFotoPrincipal = "../vistas/img/productos/".$datos["rutaProducto"].".jpg";
+						$rutaFotoPrincipal = "../views/img/categories/asesorias/".$datos["rutaAsesoria"].".jpg";
 
 						$origen = imagecreatefromjpeg($datos["fotoPrincipal"]["tmp_name"]);						
 
@@ -168,7 +398,7 @@ class ControllerProducts{
 
 						$aleatorio = mt_rand(100,999);
 
-						$rutaFotoPrincipal = "../vistas/img/productos/".$datos["rutaAsesoria"].".png";
+						$rutaFotoPrincipal = "../views/img/categories/asesorias/".$datos["rutaAsesoria"].".png";
 
 						$origen = imagecreatefrompng($datos["fotoPrincipal"]["tmp_name"]);						
 
@@ -195,7 +425,6 @@ class ControllerProducts{
 					$datosAsesoria = array(
 								   "id"=>$datos["idAsesoria"],
 								   "title"=>$datos["tituloAsesoria"],
-								   "idCategoria"=>$datos["categoria"],
 								   "hour"=>$datos["hour"],
 								   "details"=>$datos["details"],
 								   "route"=>$datos["rutaAsesoria"],
@@ -217,7 +446,6 @@ class ControllerProducts{
 					$datosAsesoria = array(
 						 		   "id"=>$datos["idAsesoria"],
 								   "title"=>$datos["tituloAsesoria"],
-								   "idCategoria"=>$datos["categoria"],
 								   "hour"=>$datos["hour"],
 								   "details"=>$datos["details"],
 								   "route"=>$datos["rutaAsesoria"],
@@ -250,13 +478,13 @@ class ControllerProducts{
 
 					swal({
 						  type: "error",
-						  title: "¡El nombre del producto no puede ir vacío o llevar caracteres especiales!",
+						  title: "¡El nombre de la asesoría no puede ir vacío o llevar caracteres especiales!",
 						  showConfirmButton: true,
 						  confirmButtonText: "Cerrar"
 						  }).then(function(result){
 							if (result.value) {
 
-							window.location = "productos";
+							window.location = "asesorias";
 
 							}
 						})
@@ -269,5 +497,74 @@ class ControllerProducts{
 
 		}
 		
+	}
+
+	/*=============================================
+	ELIMINAR ASESORIA
+	=============================================*/
+
+	public static function eliminarAsesoria(){
+
+		
+
+		if(isset($_GET["idAsesoria"])){
+
+			$datos = $_GET["idAsesoria"];
+
+			/*=============================================
+			ELIMINAR FOTO PRINCIPAL
+			=============================================*/
+
+			if($_GET["imgPrincipal"] != "" && $_GET["imgPrincipal"] != "views/img/categories/default/default.jpg"){
+
+				unlink($_GET["imgPrincipal"]);		
+
+			}
+
+			/*=============================================
+			ELIMINAR OPEN GRAPH
+			=============================================*/
+
+			if($_GET["imgPortada"] != "" && $_GET["imgPortada"] != "views/img/open_graph/default/default.jpg"){
+
+				unlink($_GET["imgPortada"]);		
+
+			}
+
+			ModelOpenGraph::eliminarOpenGraph("open_graph", $_GET["rutaOpengraph"]);
+
+			$respuesta = ModelProducts::eliminarAsesoria("products", $datos);
+
+			
+
+			if($respuesta == "ok"){
+
+				
+
+				
+
+				echo'<script>
+
+				swal({
+					  type: "success",
+					  title: "La asesoría ha sido eliminada correctamente",
+					  showConfirmButton: true,
+					  confirmButtonText: "Cerrar"
+					  }).then(function(result){
+								if (result.value) {
+
+								window.location = "asesorias";
+
+								}
+							})
+
+				</script>';
+
+			}		
+
+
+
+		}
+
 	}
 }
